[e2e] Refactor e2e tests
Update the e2e test repository to:
* Reuse an existing operatorsource for a series of test suites.
* Move e2e helper functions to the helpers package.
* Created a testsuites package which contains files that group tests
  together in test suites.
* Created a testgroups package which hosts functions that initialize
  the test environment and runs a number of test suites.
